+++ Changelog.md
@@ -8,7 +8,8 @@ Version 0.1.0 (Unreleased)
 
 Features
 --------
-- Implement Thrift-based binary injection.
+- Implement HTTP headers injection.
+- Implement thrift-based binary injection.
 - Initial library structure.
 - TraceID implementation.
 - Zipkin SpanContext implementation.